+++ js/pages/topicDetails.js
@@ -1,6 +1,7 @@
 import topicsController from '../controllers/topicsController.js';
 import storageController from '../controllers/storageController.js';
 import { createStartStudyingButton, createStudyLaterButton } from '../utils/buttonUtils.js'
+import { createTopicTitle, createTopicImage } from '../utils/topicUtils.js';
 
 export function initTopicDetails() {
     const topicId = getTopicIdFromURL();
@@ -59,23 +60,6 @@ function renderTopicDetails(topic) {
     container.append(imgWrapper, detailsWrapper);
 
     document.getElementById('topic-details').appendChild(container);
-}
-
-function createTopicTitle(topic) {
-    const title = document.createElement('h2');
-    title.className = 'card-title';
-    title.textContent = topic.title;
-    return title;
-}
-
-function createTopicImage(topic) {
-    const img = document.createElement('img');
-    img.src = topic.image;
-    img.className = 'w-100';
-    img.style.height = '300px';
-    img.style.objectFit = 'cover';
-    img.alt = topic.title;
-    return img;
 }
 
 function createTopicDescription(topic) {
